# OBPIH-7477 Fix performance issues of inventory transactions report wh…