--- a/grails-app/controllers/org/pih/warehouse/api/InventoryTransactionSummaryApiController.groovy
+++ b/grails-app/controllers/org/pih/warehouse/api/InventoryTransactionSummaryApiController.groovy
@@ -2,6 +2,7 @@ package org.pih.warehouse.api
 
 import grails.converters.JSON
 import grails.validation.ValidationException
+import org.pih.warehouse.PaginatedList
 import org.pih.warehouse.inventory.InventoryTransactionSummaryService
 import org.pih.warehouse.inventory.InventoryTransactionsSummary
 import org.pih.warehouse.report.CycleCountReportCommand
@@ -14,8 +15,8 @@ class InventoryTransactionSummaryApiController {
         if (!command.validate()) {
             throw new ValidationException("Invalid params", command.errors)
         }
-        List<InventoryTransactionsSummary> inventoryTransactions = inventoryTransactionSummaryService.getInventoryTransactionsSummary(command)
+        PaginatedList<InventoryTransactionsSummary> inventoryTransactions = inventoryTransactionSummaryService.getInventoryTransactionsSummary(command)
 
-        render([data: inventoryTransactions, totalCount: inventoryTransactions.totalCount] as JSON)
+        render([data: inventoryTransactions.list, totalCount: inventoryTransactions.totalCount] as JSON)
     }
 }