--- a/grails-app/migrations/views/changelog.xml
+++ b/grails-app/migrations/views/changelog.xml
@@ -184,12 +184,9 @@
     <sqlFile path="views/inventory-transaction-summary-helper-views-indexes.sql"/>
   </changeSet>
   <changeSet id="1748970410098-6" author="kchelstowski" runOnChange="true" runAlways="true" failOnError="true">
-    <sqlFile path="views/inventory-transactions-summary.sql"/>
-  </changeSet>
-  <changeSet id="1748970410098-7" author="kchelstowski" runOnChange="true" runAlways="true" failOnError="true">
     <sqlFile path="views/inventory-counts-helper-views.sql"/>
   </changeSet>
-  <changeSet id="1748970410098-8" author="kchelstowski" runOnChange="true" runAlways="true" failOnError="true">
+  <changeSet id="1748970410098-7" author="kchelstowski" runOnChange="true" runAlways="true" failOnError="true">
     <sqlFile path="views/inventory-counts.sql"/>
   </changeSet>
 </databaseChangeLog>