--- a/src/main/groovy/org/pih/warehouse/inventory/InventoryTransactionsSummary.groovy
+++ b/src/main/groovy/org/pih/warehouse/inventory/InventoryTransactionsSummary.groovy
@@ -5,9 +5,7 @@ import org.pih.warehouse.core.User
 import org.pih.warehouse.core.VarianceTypeCode
 import org.pih.warehouse.product.Product
 
-class InventoryTransactionsSummary implements Serializable {
-
-    String id
+class InventoryTransactionsSummary {
 
     Transaction transaction
 
@@ -26,12 +24,6 @@ class InventoryTransactionsSummary implements Serializable {
     User recordedBy
 
     Transaction baselineTransaction
-
-
-    static constraints = {
-        table "inventory_transactions_summary"
-        version false
-    }
 
     /**
      * Root causes are only captured for the cycle count
@@ -89,7 +81,6 @@ class InventoryTransactionsSummary implements Serializable {
 
     Map toJson() {
         [
-                id: id,
                 transaction: [
                     transactionNumber: transaction.transactionNumber
                 ],