… grails-app/migrations/views/inventory-transactions-summary.sql