…en reaching big amount of transactions in the system

### :sparkles: Description of Change

[//]: <> (A concise summary of what is being changed. Please provide enough context for reviewers to be able to understand the change and why it is necessary.)

**Link to GitHub issue or Jira ticket:**

**Description:**
Having reached a particular number of transactions in the system, the MariaDB query optimizer starts using different indexes and refs than it should be supposed to be using, which can cause in performance issues and scanning more rows than it should. 
When I compared the `EXPLAIN` locally and on obstg, I realized that obstg started using different transaction refs in the query than previously, and suddenly the performance dropped from ~15s on Rwinkwavu to ~4 minutes....

I decided not to fight with the automatic query optimizer anymore and force some of the indexes, and take a step back which results in:

- getting rid of the `inventory-transactions-summary` view that made the pre-filter for `facility` and `transaction_date` impossible
- instead of querying the view, we just query the same SQL statement, but with pre-filters of `facility` and `transaction_date` that cuts the rows significantly that the performance should be around ~100ms now (locally):
```sql
152 rows in set (0.077 sec)
```
vs
```sql
152 rows in set (2 min 3.716 sec)
```

If we ever need to query this logic for the whole system, we might think about bringing back the view (having assumed that we refactored the transactions not to have to have the complexed logic to handle all the edge cases), but for now, the view unfortunately was an overkill.

---
### :camera: Screenshots & Recordings (optional)

[//]: <> (If this PR contains a UI change, consider attaching one or more screenshots or recordings to help reviewers visualize the change. Otherwise, you can remove this section.)